requirements.txt
pandas
requests
pytest
mysql-connector-python
hdbscan
